backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-dateutil==2.9.0.post0